remove secondary user for slack
This causes Octoslack to fail with newer tokens. I don't know how to detect a token age, and it's clear they don't want us doing it.

This doesn't address the configuration of the alt-user, it just removes it. It's up to you on how you want to handle that.

fixes #88
see also #86